Inspect polyviz's open-data download cache

Downloads made by pv_fetch_bfs(), pv_fetch_lustat(), and pv_fetch_opendata() are kept under tools::R_user_dir("polyviz", "cache") and reused on repeat calls (pass refresh = TRUE to a fetcher to force a fresh download). pv_cache_status() lists what is currently cached; pv_cache_clear() deletes all of it.

Usage

pv_cache_status()

pv_cache_clear()

Value

pv_cache_status(): a data frame with one row per cached download (url, size_kb, fetched, file), zero rows when the cache is empty. pv_cache_clear(): the number of entries removed, invisibly.

Examples

pv_cache_status()
#>                                                                                                  url
#> 4                                                            https://www.data.lustat.ch/fa-lu-ra.csv
#> 3                                    https://dam-api.bfs.admin.ch/hub/api/dam/assets/34367751/master
#> 2         https://dam-api.bfs.admin.ch/hub/api/dam/assets/34367751/master#municipalities-g1-wgs84-v1
#> 1 https://www.uvek-gis.admin.ch/BFE/ogd/35/ogd35_schweizerische_elektrizitaetsbilanz_monatswerte.csv
#>   size_kb                  fetched
#> 4   128.1 2026-08-31 21:57:30 CEST
#> 3 10336.5 2026-08-31 21:57:39 CEST
#> 2  1667.7 2026-08-31 21:57:55 CEST
#> 1    25.2 2026-09-02 19:58:08 CEST
#>                                                                                               file
#> 4 /Users/jakestephan/Library/Caches/org.R-project.R/R/polyviz/e446824e332e80d106bbad209f62e87b.dat
#> 3 /Users/jakestephan/Library/Caches/org.R-project.R/R/polyviz/a80b1f283a383a98b335292b188140ff.dat
#> 2 /Users/jakestephan/Library/Caches/org.R-project.R/R/polyviz/960b1ce7199fde1d5b7300d0b0f306ec.dat
#> 1 /Users/jakestephan/Library/Caches/org.R-project.R/R/polyviz/11612051e57188b531d854c119c2acf6.dat