Inspect polyviz's open-data download cache

Downloads made by pv_fetch_bfs(), pv_fetch_lustat(), and pv_fetch_opendata() are kept under tools::R_user_dir("polyviz", "cache") and reused on repeat calls (pass refresh = TRUE to a fetcher to force a fresh download). pv_cache_status() lists what is currently cached; pv_cache_clear() deletes all of it.

Usage

pv_cache_status()

pv_cache_clear()

Value

pv_cache_status(): a data frame with one row per cached download (url, size_kb, fetched, file), zero rows when the cache is empty. pv_cache_clear(): the number of entries removed, invisibly.

Examples

pv_cache_status()
#>                                                                                                   url
#> 16                                                            https://www.data.lustat.ch/fa-lu-ra.csv
#> 11                                    https://dam-api.bfs.admin.ch/hub/api/dam/assets/34367751/master
#> 8          https://dam-api.bfs.admin.ch/hub/api/dam/assets/34367751/master#municipalities-g1-wgs84-v1
#> 4  https://www.uvek-gis.admin.ch/BFE/ogd/35/ogd35_schweizerische_elektrizitaetsbilanz_monatswerte.csv
#> 5                                           https://www.data.lustat.ch/szbv-lu-2025-2055-referenz.csv
#> 6                                                          https://www.data.lustat.ch/gefis-lu-jr.csv
#> 9                                                    https://www.data.lustat.ch/szbd-lu-2025-2039.csv
#> 10                                                         https://www.data.lustat.ch/grwahlen-lu.csv
#> 13                                              https://www.data.lustat.ch/szbv-lu-2025-2055-hoch.csv
#> 17                                              https://www.data.lustat.ch/szbv-lu-2025-2055-tief.csv
#> 15    https://disseminate.stats.swiss/rest/dataflow/all/all/latest?format=fusion-json&detail=allstubs
#> 1                        https://disseminate.stats.swiss/rest/dataflow/all/all/latest?detail=allstubs
#> 7                                                            https://www.data.lustat.ch/fa-lu-tla.csv
#> 12                                                           https://www.data.lustat.ch/fa-lu-sla.csv
#> 14                                                           https://www.data.lustat.ch/fa-lu-ila.csv
#> 2   https://disseminate.stats.swiss/rest/data/CH1.STATPOP,DF_STATPOP_PHH/all?format=csvfilewithlabels
#> 3       https://disseminate.stats.swiss/rest/data/CH1.SSV,DF_SSV_POP_BIL/all?format=csvfilewithlabels
#>     size_kb                  fetched
#> 16    128.1 2026-08-31 21:57:30 CEST
#> 11  10336.5 2026-08-31 21:57:39 CEST
#> 8    1667.7 2026-08-31 21:57:55 CEST
#> 4      25.2 2026-09-02 19:58:08 CEST
#> 5  113126.4  2026-09-07 13:02:14 BST
#> 6   65013.2  2026-09-07 13:04:46 BST
#> 9     126.8  2026-09-07 13:04:51 BST
#> 10     78.4  2026-09-07 13:04:51 BST
#> 13 109258.3  2026-09-07 13:06:09 BST
#> 17 109254.5  2026-09-07 13:12:40 BST
#> 15    228.7  2026-09-07 16:01:03 BST
#> 1     228.7  2026-09-07 16:01:11 BST
#> 7     128.3  2026-09-07 16:04:11 BST
#> 12     99.5  2026-09-07 16:04:12 BST
#> 14     93.0  2026-09-07 16:04:12 BST
#> 2   50661.1  2026-09-07 16:04:22 BST
#> 3    2518.3  2026-09-07 16:04:25 BST
#>                                                                                                file
#> 16 /Users/jakestephan/Library/Caches/org.R-project.R/R/polyviz/e446824e332e80d106bbad209f62e87b.dat
#> 11 /Users/jakestephan/Library/Caches/org.R-project.R/R/polyviz/a80b1f283a383a98b335292b188140ff.dat
#> 8  /Users/jakestephan/Library/Caches/org.R-project.R/R/polyviz/960b1ce7199fde1d5b7300d0b0f306ec.dat
#> 4  /Users/jakestephan/Library/Caches/org.R-project.R/R/polyviz/11612051e57188b531d854c119c2acf6.dat
#> 5  /Users/jakestephan/Library/Caches/org.R-project.R/R/polyviz/260b5896162d8a40565f451f8201cebb.dat
#> 6  /Users/jakestephan/Library/Caches/org.R-project.R/R/polyviz/3f7cddd19ac1b22f9431276fcebf9ce5.dat
#> 9  /Users/jakestephan/Library/Caches/org.R-project.R/R/polyviz/a193304b3e894d9c7733b4b3c8536b88.dat
#> 10 /Users/jakestephan/Library/Caches/org.R-project.R/R/polyviz/a6cdd6442bbee6ebe411e560274be095.dat
#> 13 /Users/jakestephan/Library/Caches/org.R-project.R/R/polyviz/aa645130b95a2f1b0bf2336730b27288.dat
#> 17 /Users/jakestephan/Library/Caches/org.R-project.R/R/polyviz/f572100539eb1892e0c1a671f57aba58.dat
#> 15 /Users/jakestephan/Library/Caches/org.R-project.R/R/polyviz/cec783b0f17b7e4b9664b4c3fd9a05c1.dat
#> 1  /Users/jakestephan/Library/Caches/org.R-project.R/R/polyviz/058241473b2a253d5835257099686e1d.dat
#> 7  /Users/jakestephan/Library/Caches/org.R-project.R/R/polyviz/4214ab1da304e6346186a25ad073849f.dat
#> 12 /Users/jakestephan/Library/Caches/org.R-project.R/R/polyviz/a8e7b54bd5433894513122d1ee620b75.dat
#> 14 /Users/jakestephan/Library/Caches/org.R-project.R/R/polyviz/b4e477e9bfa2fe6074ee6259ac5cd507.dat
#> 2  /Users/jakestephan/Library/Caches/org.R-project.R/R/polyviz/058e7a9b3262f99f0a670b88f42c7919.dat
#> 3  /Users/jakestephan/Library/Caches/org.R-project.R/R/polyviz/0e75f3296998612e975d7092e816f10d.dat